Create basic25.c
basic25.c: The main attraction of this program was to remind me to
always pay attention to every detail, even if the problem seems easy.
For example, when I was to print the month, I forgot that arrays are
indexed from 0 in C instead of 1, so I had to deduce 1 from the user
submitted integer. I should have noticed it when I made the draft of
this program, in the pseudocode stage.