Create basic35.c and basic36.c
basic35.c: This was easy. It can be forked from the code of basic34.c.

basic36.c: This was difficult because I did not read the documentation
of the string functions properly so I improvised instead of a systematic
debugging approach. I am not completely satified with the result,
because this program compares chunks of the password given by the user
in 5-byte-long chunks, instead of only one comparison. Nevertheless, if
the correct password is typed in, it works.