Create basic37.c
basic37.c: It was more difficult than I thought. I should have checked
the pseudocode for errors or even do some manual evaluations before
writing everything down. Of course the problems arose around edge cases
because the boundaries of the quadrants are not defined and not
specified in the exercise.