Create basic45.c and basic46.c
basic45.c: I figured out that the for loop is too complicated because
the index variable 'k' produces an integer division. So instead of
creating yet another varialbe of the type double, I used a while loop
with manual increments on a type double variable.

basic46.c: I did not find the formula for this exercise as the first and
last term cannot be written as (2^k +_1)/2^k. So I hard-coded the sum.